requirements.txt
# Core dependencies
yfinance>=0.2.18
requests>=2.31.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
watchdog>=3.0.0

# Database and storage
aiosqlite>=0.19.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Optional enhancements for future phases
# redis>=5.0.0
# influxdb-client>=1.38.0
# psycopg2-binary>=2.9.0